Need to figure out how best way to associate courses to a term in progress.

+++ app/src/main/java/com/humphrey/c196/DAO/CourseDAO.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface CourseDAO {
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    long insert(Course course);
+    int insert(Course course);
 
     @Update
     void update(Course course);
